# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Argumentation Analysis API",
    "description": "API d'analyse argumentative multi-agents avec intégration Java/Tweety, composants Lego (CapabilityRegistry), et pipelines unifiés. Supporte l'analyse Dung, Toulmin, ASPIC+, détection de sophismes, et orchestration de débats.",
    "version": "2.1.0",
    "x-integration-status": {
      "registered_capabilities": 25,
      "integrated_projects": 10,
      "architecture": "Lego (CapabilityRegistry + Semantic Kernel plugins)"
    }
  },
  "paths": {
...
    "/jtms/sessions": {
      "post": {
        "tags": [
          "JTMS"
        ],
        "summary": "Create Session",
        "description": "Crée une nouvelle session de travail pour un agent, qui peut contenir\nplusieurs instances JTMS.",
        "operationId": "create_session_jtms_sessions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "JTMS"
        ],
        "summary": "List Sessions",
        "description": "Liste toutes les sessions existantes, avec la possibilité de filtrer par\nagent_id et par statut.",
        "operationId": "list_sessions_jtms_sessions_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSessionRequest": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id",
            "description": "ID de l'agent"
          },
          "session_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Name",
            "description": "Nom descriptif de la session"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Métadonnées additionnelles",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "agent_id"
        ],
        "title": "CreateSessionRequest",
        "description": "Requête de création de session."
      },
      "CreateSessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "ID de la session créée"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id",
            "description": "ID de l'agent"
          },
          "session_name": {
            "type": "string",
            "title": "Session Name",
            "description": "Nom de la session"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "Date de création"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Statut de l'opération",
            "default": "success"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "agent_id",
          "session_name",
          "created_at"
        ],
        "title": "CreateSessionResponse",
        "description": "Réponse de création de session."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionInfo": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "ID de la session"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id",
            "description": "ID de l'agent propriétaire"
          },
          "session_name": {
            "type": "string",
            "title": "Session Name",
            "description": "Nom de la session"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "Date de création"
          },
          "last_accessed": {
            "type": "string",
            "title": "Last Accessed",
            "description": "Dernier accès"
          },
          "checkpoint_count": {
            "type": "integer",
            "title": "Checkpoint Count",
            "description": "Nombre de checkpoints",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "agent_id",
          "session_name",
          "created_at",
          "last_accessed"
        ],
        "title": "SessionInfo",
        "description": "Informations sur une session."
      },
      "SessionListResponse": {
        "properties": {
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/SessionInfo"
            },
            "type": "array",
            "title": "Sessions",
            "description": "Liste des sessions",
            "default": []
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "description": "Nombre total de sessions",
            "default": 0
          },
          "agent_filter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Filter",
            "description": "Filtre agent appliqué"
          },
          "status_filter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Filter",
            "description": "Filtre statut appliqué"
          }
        },
        "type": "object",
        "title": "SessionListResponse",
        "description": "Réponse de liste de sessions."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ]
}